ADD CLASS PARSERS THAT CONTAINS PARSER FUNCTIONS #77
