+++ app/src/main/java/ch/epfl/sweng/zuluzulu/URLTools/AssociationsUrlHandler.java
@@ -1,20 +1,14 @@
-package ch.epfl.sweng.zuluzulu.Structure;
+package ch.epfl.sweng.zuluzulu.URLTools;
 
 import android.os.AsyncTask;
 import android.util.Log;
 
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.UnsupportedEncodingException;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import io.opencensus.common.Function;
 
@@ -29,6 +23,7 @@ public class AssociationsUrlHandler<T> extends AsyncTask<String, Void, T> {
 
     /**
      * Create a new AssociationUrlHandler
+     *
      * @param listener The callback function that will be use on PostExecute
      */
     public AssociationsUrlHandler(Function<T, Void> listener, Function<BufferedReader, T> parser) {
@@ -53,6 +48,7 @@ public class AssociationsUrlHandler<T> extends AsyncTask<String, Void, T> {
 
     /**
      * Connect to the URL, check if the response code is OK (200)
+     *
      * @param url Url
      * @return HttpURLConnection or null of response is not OK
      * @throws IOException Throw exception if it cannot connect
@@ -81,6 +77,7 @@ public class AssociationsUrlHandler<T> extends AsyncTask<String, Void, T> {
 
     /**
      * Connect to the URL, parse it and return the values found
+     *
      * @param url The URL
      * @return T Return object of type T with all the values founded
      */
@@ -102,7 +99,6 @@ public class AssociationsUrlHandler<T> extends AsyncTask<String, Void, T> {
             Log.d(TAG, "null UrlConnection");
             return null;
         }
-
 
 
         BufferedReader in = null;
@@ -131,53 +127,6 @@ public class AssociationsUrlHandler<T> extends AsyncTask<String, Void, T> {
         return datas;
     }
 
-    /**
-     * This function will parse the datas and return a arraylist of strings
-     *
-     * @param in Input stream
-     * @return ArrayList of strings , values separated by a comma
-     */
-    public static ArrayList<String> parseAssociationsData(BufferedReader in) {
-        // regex
-        Pattern p = Pattern.compile("&#\\d+.* <a href=\"(.*?)\".*>(.*)</a>.*\\((.+)\\)<.*br />.*");
-
-        ArrayList<String> results = new ArrayList<>();
-
-        String inputLine;
-
-        try {
-            while ((inputLine = in.readLine()) != null) {
-                Matcher m = p.matcher(inputLine);
-                if (m.find()) {
-                    // remove the span tag
-                    String description = m.group(3).replaceAll("<.*?>", "");
-
-                    StringBuilder sb = new StringBuilder();
-                    sb.append(m.group(1));
-                    sb.append(',');
-                    sb.append(m.group(2));
-                    sb.append(',');
-                    sb.append(description);
-
-                    // remplace html unicode to char
-                    String result = sb.toString()
-                            .replaceAll("&#8217;", "'")
-                            .replaceAll("&#8211;", "-")
-                            .replaceAll("&gt;", ">")
-                            .replaceAll("&amp;", "&");
-
-                    results.add(result);
-                }
-            }
-            in.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-            Log.d(TAG, "Could not parse datas");
-            return null;
-        }
-
-        return results;
-    }
 
     /**
      * Open the url
